change attack and defense from i32 to f32
This will grant finer-grained control over these stats in the future.

+++ src/components.rs
@@ -15,15 +15,15 @@ pub struct Asleep {
 pub struct BlocksTile;
 
 pub struct CombatBonus {
-    pub attack: i32,
-    pub defense: i32,
+    pub attack: f32,
+    pub defense: f32,
 }
 
 pub struct CombatStats {
     pub max_hp: i32,
     pub hp: i32,
-    pub attack: i32,
-    pub defense: i32,
+    pub attack: f32,
+    pub defense: f32,
 }
 
 pub struct Consumable;
